file apis implement

+++ api/src/main/java/com/ke/bella/files/api/FileController.java
@@ -4,9 +4,11 @@ import static com.ke.bella.files.configuration.Configs.MAX_SIZE_IN_MB;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.List;
 
 import javax.servlet.http.HttpServletResponse;
 
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpHeaders;
@@ -22,7 +24,9 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
 import com.ke.bella.files.annotations.FileAPI;
+import com.ke.bella.files.protocol.FileException.FileNotFoundException;
 import com.ke.bella.files.protocol.FileException.FileTooLargeException;
+import com.ke.bella.files.protocol.FileException.ProgressNotFoundException;
 import com.ke.bella.files.protocol.FileUrl;
 import com.ke.bella.files.protocol.OpenAIFile;
 import com.ke.bella.files.protocol.OpenapiListResponse;
@@ -59,7 +63,7 @@ public class FileController {
         File tmpFile = null;
         try {
             String extension = FileExtensionDetectUtil.detectExtension(file.getOriginalFilename());
-            String suffix = extension == null ? "" : "." + extension;
+            String suffix = StringUtils.isEmpty(extension) ? "" : "." + extension;
             File tmpDir = new File(tmpFileDir);
             tmpFile = File.createTempFile("tmp", suffix, tmpDir);
             file.transferTo(tmpFile);
@@ -80,23 +84,62 @@ public class FileController {
             @RequestParam(value = "limit", required = false, defaultValue = "10000") Integer limit,
             @RequestParam(value = "order", required = false, defaultValue = "desc") String order,
             @RequestParam(value = "after", required = false) String after) {
-        return fileService.list(purpose, limit, order, after);
+        if(limit < 1) {
+            throw new IllegalArgumentException(limit + " is less than the minimum of 1");
+        }
+        if(limit > 10000) {
+            throw new IllegalArgumentException(limit + " is greater than the maximum of 10000");
+        }
+        if(!order.equals("desc") && !order.equals("asc")) {
+            throw new IllegalArgumentException(order + "is not one of ['asc', 'desc']");
+        }
+
+        // 如果after为空，说明用户没有指定after，此时查询所有数据
+        // 如果传了一个无效after，才抛出错误
+        if(!StringUtils.isEmpty(after) && fileService.getFile(after) == null) {
+            throw new IllegalArgumentException("Invalid after: " + after);
+        }
+
+        List<OpenAIFile> files = fileService.list(purpose, limit + 1, order, after);
+        OpenapiListResponse<OpenAIFile> res = new OpenapiListResponse<>();
+        if(files.size() > limit) {
+            files = files.subList(0, limit);
+            res.setHasMore(true);
+            res.setLastId(files.get(limit - 1).getId());
+        } else if(!files.isEmpty()) {
+            res.setLastId(files.get(files.size() - 1).getId());
+        }
+        res.setData(files);
+        return res;
     }
 
     @GetMapping("/{file_id}")
     public OpenAIFile get(@PathVariable("file_id") String fileId) {
-        return fileService.get(fileId);
+        OpenAIFile file = fileService.getFile(fileId);
+        if(file == null) {
+            throw new FileNotFoundException(fileId);
+        }
+        return file;
     }
 
     @DeleteMapping("/{file_id}")
     public OpenAIFile delete(@PathVariable("file_id") String fileId) {
-        return fileService.delete(fileId);
+        OpenAIFile file = fileService.getFile(fileId);
+        if(file == null) {
+            throw new FileNotFoundException(fileId);
+        }
+        fileService.delete(fileId);
+        return file;
     }
 
     @GetMapping("/{file_id}/content")
     public void retrieveContentRedirect(
             HttpServletResponse response,
             @PathVariable("file_id") String fileId) {
+        OpenAIFile file = fileService.getFile(fileId);
+        if(file == null) {
+            throw new FileNotFoundException(fileId);
+        }
         FileUrl fileUrl = fileService.getUrl(fileId);
         String redirectUrl = fileUrl.getS3Url();
         response.setHeader(HttpHeaders.LOCATION, redirectUrl);
@@ -107,19 +150,36 @@ public class FileController {
     public FileUrl getUrl(
             @PathVariable("file_id") String fileId,
             @RequestParam(value = "expires", required = false, defaultValue = "86400") Long expires) {
+        OpenAIFile file = fileService.getFile(fileId);
+        if(file == null) {
+            throw new FileNotFoundException(fileId);
+        }
         return fileService.getUrl(fileId, expires);
     }
 
     @PostMapping("{file_id}/progress/{progress_name}")
-    public Progress updateProgress(@RequestBody UpdateProgressRequestData data,
+    public Progress updateProgress(
+            @RequestBody UpdateProgressRequestData data,
             @PathVariable("file_id") String fileId,
             @PathVariable("progress_name") String progressName) {
-        return fileService.updateProgress(data, fileId, progressName);
+        if(fileService.getFile(fileId) == null) {
+            throw new IllegalArgumentException("Invalid fileId: " + fileId);
+        }
+        fileService.updateProgress(data, fileId, progressName);
+        return fileService.getProgress(fileId, progressName);
     }
 
     @GetMapping("/{file_id}/progress")
-    public Progress getProgress(@PathVariable("file_id") String fileId,
+    public Progress getProgress(
+            @PathVariable("file_id") String fileId,
             @RequestParam("progress_name") String progressName) {
-        return fileService.getProgress(fileId, progressName);
+        if(fileService.getFile(fileId) == null) {
+            throw new IllegalArgumentException("Invalid fileId: " + fileId);
+        }
+        Progress res = fileService.getProgress(fileId, progressName);
+        if(res == null) {
+            throw new ProgressNotFoundException(fileId, progressName);
+        }
+        return res;
     }
 }
